pyproject.toml
[project]
name = "langchain-chatbot-tutorial"
version = "0.1.0"
description = ""
authors = [
    {name = "Tri Dang",email = "quangtrister@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain-anthropic (>=0.3.10,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "ipython (>=9.1.0,<10.0.0)",
    "langgraph (>=0.3.30,<0.4.0)",
    "tavily-python (>=0.5.4,<0.6.0)",
    "numpy (==2.2.0)",
    "langchain-community (>=0.3.21,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
